Update ytd and mtd calculation
Fixed bug in ytd and mtd calculation. ytd and mtd are not calculated for short return series. self.mtd is only assigned something other than total_return, if more than one monthly return exists - which is unnecessary. Same issue for ytd - but with a stricter condition of more data before it is calculated. One daily return is enough to calculate ytd and mtd returns.